Add Dockerfile to build image
Add a `Dockerfile` and `.dockerignore` to be able to build a Docker
image for the application as an example of how to build a small image
using a multi-stage build.

Also add a job to the check workflow that tests that the `Dockerfile`
builds and image can be run after.